+++ packages/multichain-api-middleware/src/handlers/wallet-getSession.ts
@@ -7,13 +7,30 @@ import {
   Caip25EndowmentPermissionName,
   getSessionScopes,
 } from '@metamask/chain-agnostic-permission';
-import type { Caveat } from '@metamask/permission-controller';
-import type { CaipAccountId } from '@metamask/utils';
 import type {
+  JsonRpcEngineEndCallback,
+  JsonRpcEngineNextCallback,
+  MethodHandler,
+} from '@metamask/json-rpc-engine';
+import type { Caveat } from '@metamask/permission-controller';
+import type {
+  CaipAccountId,
   CaipChainId,
+  JsonRpcParams,
   JsonRpcRequest,
-  JsonRpcSuccess,
+  PendingJsonRpcResponse,
 } from '@metamask/utils';
+
+type WalletGetSessionResult = { sessionScopes: NormalizedScopesObject };
+
+type WalletGetSessionHooks = {
+  getCaveatForOrigin: (
+    endowmentPermissionName: string,
+    caveatType: string,
+  ) => Caveat<typeof Caip25CaveatType, Caip25CaveatValue>;
+  getNonEvmSupportedMethods: (scope: CaipChainId) => string[];
+  sortAccountIdsByLastSelected: (accounts: CaipAccountId[]) => CaipAccountId[];
+};
 
 /**
  * Handler for the `wallet_getSession` RPC method as specified by [CAIP-312](https://chainagnostic.org/CAIPs/caip-312).
@@ -31,21 +48,12 @@ import type {
  * @param hooks.sortAccountIdsByLastSelected - A function that accepts an array of CaipAccountId and returns an array of CaipAccountId sorted by corresponding last selected account in the wallet.
  * @returns Nothing.
  */
-async function walletGetSessionHandler(
+async function handleWalletGetSession(
   _request: JsonRpcRequest & { origin: string },
-  response: JsonRpcSuccess<{ sessionScopes: NormalizedScopesObject }>,
-  _next: () => void,
-  end: () => void,
-  hooks: {
-    getCaveatForOrigin: (
-      endowmentPermissionName: string,
-      caveatType: string,
-    ) => Caveat<typeof Caip25CaveatType, Caip25CaveatValue>;
-    getNonEvmSupportedMethods: (scope: CaipChainId) => string[];
-    sortAccountIdsByLastSelected: (
-      accounts: CaipAccountId[],
-    ) => CaipAccountId[];
-  },
+  response: PendingJsonRpcResponse<WalletGetSessionResult>,
+  _next: JsonRpcEngineNextCallback,
+  end: JsonRpcEngineEndCallback,
+  hooks: WalletGetSessionHooks,
 ) {
   let caveat;
   try {
@@ -71,12 +79,23 @@ async function walletGetSessionHandler(
   return end();
 }
 
+type WalletGetSessionMethodHandler = MethodHandler<
+  WalletGetSessionHooks,
+  never,
+  JsonRpcParams,
+  WalletGetSessionResult,
+  { origin: string }
+>;
+
 export const walletGetSession = {
-  methodNames: ['wallet_getSession'],
-  implementation: walletGetSessionHandler,
+  implementation: handleWalletGetSession,
   hookNames: {
     getCaveatForOrigin: true,
     getNonEvmSupportedMethods: true,
     sortAccountIdsByLastSelected: true,
   },
+} satisfies WalletGetSessionMethodHandler;
+
+export const walletGetSessionHandler = {
+  wallet_getSession: walletGetSession,
 };